+++ db/blob/blob_log_format.h
@@ -107,7 +107,7 @@ struct BlobLogRecord {
   // Note that the offset field of BlobIndex actually points to the blob value
   // as opposed to the start of the blob record. The following method can
   // be used to calculate the adjustment needed to read the blob CRC field.
-  static size_t CalculateAdjustmentForRecordHeader(size_t key_size);
+  static uint64_t CalculateAdjustmentForRecordHeader(uint64_t key_size);
 
   uint64_t key_size = 0;
   uint64_t value_size = 0;
@@ -129,7 +129,7 @@ struct BlobLogRecord {
 };
 
 // Checks whether a blob offset is potentially valid or not.
-inline bool IsValidBlobOffset(uint64_t value_offset, size_t key_size) {
+inline bool IsValidBlobOffset(uint64_t value_offset, uint64_t key_size) {
   return value_offset >=
          BlobLogHeader::kSize + BlobLogRecord::kHeaderSize + key_size;
 }
